refacto(sub): move to sub package
